Extract document area layout manager
Refactors document workspace layout concerns out of `DocumentSectionContainer` into a new `DocumentAreaLayout` class. The new type now owns area visibility, split state/ratios, splitters, chrome, toolbar placement, snapping, and layout reset, while `DocumentSectionContainer` focuses on tab/document ownership and active-document behavior.

Updated `WorkspacePanel` and tab-menu flows to route layout operations through `SectionContainer.Areas`, preserving existing behavior with cleaner separation of responsibilities. Also fixes `WindowStateHelper` minimum-size logging to report `WindowSizeScale` and clearer "scale" wording.
